Made slider have visible button state
The absence of the slider handle makes sliders and progress bars look identical. The glass bubble was made visible always so they could be differentiated. The correct state is that it has a visible button, but can morph to the glass bubble, hiding the visible button, which is not currently available on plasma.