Use slider with outline in doncsugar
Plasma has the slider assets for the audio plasmoid to use the fill bar stacked with transparency to create the audio feedback indication. However, when the fill bar is opaque, it is impossible to see the feedback. Adding an outline slightly provides feedback, but is not ideal. Allowing 2 separate assets for the feedback is the correct solution.