cleaned up tabs
The round tab groove, while aesthetically pleasing, is visually incorrect when the number of tabs exceeds what the tabbar can support and it needs to extend outside of it. Since the groove has round corners, it wrongly appears like the end is on screen. The best way to achieve the style is with the legacy technique used in Sumac, but the inability to have the tabs fully move under the active tab reduces the effectiveness of the style. The inactive assets still need to be updated.