updater slider
The glass bubble styling could not be used because Kvantum is surprisingly, more limited than Plasma. In particular, the styling of the slider in the plasma theme cannot be achieved because Kvantum draws the fill and groove very sharply, splitting them on a perfect line instead of how the plasma themes do it with a bar that is drawn and expanded to meet the current percentage.